test route with fetch options
